+++ src/pages/blind-match/components/agreement/agreement.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+export const agreementoption = style({
+  display: 'flex',
+  height: '5rem',
+  marginTop: '3.4rem',
+});
