+++ src/pages/blind-match/components/agreement/agreement.tsx
@@ -1,0 +1,24 @@
+import { BLIND_MATCH_TEXT } from '@pages/blind-match/constants/blind-match-text';
+
+import AgreementOption from '@shared/components/agreement-option/agreement-option';
+
+import * as styles from './agreement.css';
+
+interface AgreementProps {
+  checked: boolean;
+  onChange: (isAgreed: boolean) => void;
+}
+
+const Agreement = ({ checked, onChange }: AgreementProps) => {
+  return (
+    <div className={styles.agreementoption}>
+      <AgreementOption
+        label={BLIND_MATCH_TEXT.CONSENT}
+        navigateTo=""
+        checked={checked}
+        onChange={onChange}
+      />
+    </div>
+  );
+};
+export default Agreement;
