+++ src/pages/blind-match/components/entrypage/entry.css.ts
@@ -2,30 +2,10 @@ import { style } from '@vanilla-extract/css';
 
 import { themeVars } from '@shared/styles';
 
-export const title = style({
-  padding: '1.3rem 2.4rem 1rem 2.4rem',
-});
-
 export const container = style({
   display: 'grid',
   justifyContent: 'center',
   alignItems: 'center',
-});
-
-export const textcontainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-  height: '11.3rem',
-});
-export const message = style({
-  ...themeVars.fontStyles.title_b_16,
-  color: themeVars.color.main_blue,
-});
-export const time = style({
-  ...themeVars.fontStyles.button_r_14,
-  color: themeVars.color.gray_900,
 });
 
 export const attention = style({
@@ -41,12 +21,6 @@ export const attention = style({
 });
 
 export const error = style({ color: themeVars.color.error_red });
-
-export const agreementoption = style({
-  display: 'flex',
-  height: '5rem',
-  marginTop: '3.4rem',
-});
 
 export const notice = style({
   display: 'flex',
