+++ src/pages/blind-match/components/entrypage/entry.tsx
@@ -1,13 +1,14 @@
 import { useState } from 'react';
 
-import AgreementOption from '@shared/components/agreement-option/agreement-option.tsx';
 import Button from '@shared/components/button/button';
-import Title from '@shared/components/title/title';
 import { IcSvgCaution } from '@shared/icons';
 
 import * as styles from './entry.css.ts';
 import { BLIND_MATCH_TEXT } from '../../constants/blind-match-text';
+import Agreement from '../agreement/agreement.tsx';
+import Message from '../message/message.tsx';
 import ConfirmModal from '../modal/confirm-modal.tsx';
+import EntryTitle from '../title/title.tsx';
 import UseInfoForm from '../use-info-form/use-info-form';
 
 interface EntryPageProps {
@@ -48,24 +49,14 @@ const EntryPage = ({ currentDay }: EntryPageProps) => {
   const handleCloseModal = () => {
     setIsModalOpen(false);
   };
+
   return (
     <>
-      <div className={styles.title}>
-        <Title
-          mainTitle={BLIND_MATCH_TEXT.TITLE}
-          subTitle={BLIND_MATCH_TEXT.SUBTITLE}
-        />
-      </div>
+      <EntryTitle />
+
+      <Message currentDay={currentDay} />
 
       <div className={styles.container}>
-        <div className={styles.textcontainer}>
-          <p className={styles.message}>
-            {currentDay}
-            {BLIND_MATCH_TEXT.MESSAGE}
-          </p>
-          <p className={styles.time}>{BLIND_MATCH_TEXT.TIME}</p>
-        </div>
-
         <UseInfoForm
           instaId={instaId}
           phoneNumber={phoneNumber}
@@ -96,14 +87,10 @@ const EntryPage = ({ currentDay }: EntryPageProps) => {
           {BLIND_MATCH_TEXT.ATTENTION}
         </div>
 
-        <div className={styles.agreementoption}>
-          <AgreementOption
-            label={BLIND_MATCH_TEXT.CONSENT}
-            navigateTo=""
-            checked={agreed}
-            onChange={handleConsentChange}
-          />
-        </div>
+        <Agreement
+          checked={agreed}
+          onChange={handleConsentChange}
+        />
 
         <Button onClick={handleApplyClick}>번호팅 신청하기</Button>
