+++ src/pages/blind-match/components/title/title.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const title = style({
+  padding: '1.3rem 2.4rem 1rem 2.4rem',
+});
